requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic==2.10.4
joblib==1.4.2
pandas==2.2.3
numpy==2.0.2
shap==0.46.0
xgboost==2.1.3
scikit-learn==1.6.1
python-dotenv==1.0.1
